test/requirements.txt
boto3>=1.26.0
kafka-python>=2.0.2
python-dateutil>=2.8.2 